+++ js/index.js
@@ -1,3 +1,4 @@
 export * from "./base";
 export * from "./invoice";
+export * from "./sequence";
 export * from "./tax";