+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-*
+* Method to list all sequences - [peri-invoicing/#25](https://github.com/ripe-tech/peri-invoicing/issues/25)
+* Method to create sequence - [peri-invoicing/#25](https://github.com/ripe-tech/peri-invoicing/issues/25)
 
 ### Changed